refactor(time): use base package `me.letsdev.time`
